+++ src/lib.rs
@@ -431,8 +431,6 @@ mod ser;
 pub mod value;
 
 mod io;
-#[cfg(feature = "std")]
-mod iter;
 #[cfg(feature = "float_roundtrip")]
 mod lexical;
 mod number;